Generate Multivariate Normal Data \( \mathbf{X} \sim \mathcal{N}_{k} \left( \boldsymbol{\mu}, \boldsymbol{\Sigma} \right) %(\#eq:dist-X-mvn) \) from the Reticular Action Model (RAM) Matrices

Generates an \(n \times k\) multivariate data matrix or a list of \(n \times k\) multivariate data matrices of length R from the multivariate normal distribution $$ \mathbf{X} \sim \mathcal{N}_{k} \left( \boldsymbol{\mu}, \boldsymbol{\Sigma} \right) %(\#eq:dist-X-mvn) $$ The model-implied matrices used to generate data is derived from the Reticular Action Model (RAM) Matrices.

mvnram(
  n,
  A,
  S,
  F,
  I,
  M = NULL,
  mu = NULL,
  tol = 1e-06,
  empirical = FALSE,
  R = NULL,
  par = FALSE,
  ncores = NULL,
  mc = TRUE,
  lb = FALSE,
  cl_eval = FALSE,
  cl_export = FALSE,
  cl_expr,
  cl_vars
)

Arguments

n
Integer. Sample size.

A
\(\mathbf{A}\) matrix. Asymmetric paths (single-headed arrows), such as regression coefficients and factor loadings.

S
\(\mathbf{S}\) matrix. Symmetric paths (double-headed arrows), such as variances and covariances.

F
\(\mathbf{F}\) matrix. Filter matrix used to select the observed variables.

I
\(\mathbf{I}\) matrix. Identity matrix.

M
Numeric vector. Mean structure. Vector of means and intercepts.

mu
Numeric vector. Location parameter mean vector \(\boldsymbol{\mu}\) of length \(k\).

tol
Numeric. Tolerance (relative to largest variance) for numerical lack of positive-definiteness in Sigma.

empirical
Logical. If TRUE, mu and Sigma specify the empirical not population mean and covariance matrix.

R
Integer. Number of Monte Carlo replications. If R is not provided, the function produces a single random data set. If R is an integer greater than 1, (e.g., R = 10), the function produces multiple random data sets stored in each element of a list of length R. par and all succeeding arguments are only relevant when R > 1.

par
Logical. If TRUE, use multiple cores. If FALSE, use lapply().

ncores
Integer. Number of cores to use if par = TRUE. If unspecified, defaults to detectCores() - 1.

mc
Logical. If TRUE, use parallel::mclapply(). If FALSE, use parallel::parLapply() or parallel::parLapplyLB(). Ignored if par = FALSE.

lb
Logical. If TRUE use parallel::parLapplyLB(). If FALSE, use parallel::parLapply(). Ignored if par = FALSE and mc = TRUE.

cl_eval
Logical. Execute parallel::clusterEvalQ() using cl_expr. Ignored if mc = TRUE.

cl_export
Logical. Execute parallel::clusterExport() using cl_vars. Ignored if mc = TRUE.

cl_expr
Expression. Expression passed to parallel::clusterEvalQ() Ignored if mc = TRUE.

cl_vars
Character vector. Names of objects to pass to parallel::clusterExport() Ignored if mc = TRUE.

Value

If R = NULL or R = 1, returns an \(n \times k\) multivariate normal data matrix generated using the provided or derived mean vector mu \(\left( \boldsymbol{\mu} \right)\) and the derived variance-covariance matrix Sigma \(\left( \boldsymbol{\Sigma} \right)\). If R is an integer greater than 1, (e.g., R = 10) returns a list of length R of \(n \times k\) multivariate normal data matrices generated using the provided or derived mean vector mu and the derived variance-covariance matrix Sigma.

Details

The multivariate normal distribution has two parameters, namely the \(k \times 1\) mean vector mu \(\left( \boldsymbol{\mu} \right)\) and the \(k \times k\) variance-covariance matrix Sigma \(\left( \boldsymbol{\Sigma} \right)\).

The mean vector mu \(\left( \boldsymbol{\mu} \right)\) can be supplied directly using the mu argument. It can also be derived using M. Note that the argument mu \(\left( \boldsymbol{\mu} \right)\) takes precedence over M If mu is not provided, it is computed using M with the jeksterslabRsem::ram_mutheta() function. If both mu and M are not provided, mu is set to a vector of zeroes with the appropriate length.

The variance-covariance matrix Sigma \(\left(\boldsymbol{\Sigma}\right)\) is derived from the RAM matrices A, S, F, and I.

mu \(\left( \boldsymbol{\mu} \right)\) and Sigma \(\left(\boldsymbol{\Sigma}\right)\) are then used by the mvn() to generate multivariate normal data.

Options for explicit parallelism are provided when R > 1 especially when R is large. See par and suceeding arguments.

References

McArdle, J. J. (2013). The development of the RAM rules for latent variable structural equation modeling. In A. Maydeu-Olivares & J. J. McArdle (Eds.), Contemporary Psychometrics: A festschrift for Roderick P. McDonald (pp. 225--273). Lawrence Erlbaum Associates.

McArdle, J. J., & McDonald, R. P. (1984). Some algebraic properties of the Reticular Action Model for moment structures. British Journal of Mathematical and Statistical Psychology, 37 (2), 234--251.

See also

Examples

A <- matrix(
  data = c(
    0, 0.26^(1 / 2), 0,
    0, 0, 0.26^(1 / 2),
    0, 0, 0
  ),
  ncol = 3
)
S <- F <- I <- diag(3)
S[1, 1] <- 225
S[2, 2] <- 166.5
S[3, 3] <- 166.5
mu <- c(100, 100, 100)
# single random data set------------------------------------------------------
X <- mvnram(
  n = 100,
  A = A,
  S = S,
  F = F,
  I = I,
  mu = mu
)
str(X, list.len = 6)

#> 'data.frame':	100 obs. of  3 variables:
#>  $ V1: num  109.5 105.3 117.6 85.7 97.5 ...
#>  $ V2: num  92.3 128.5 107.6 70.2 107 ...
#>  $ V3: num  93.6 80.4 111.8 85.8 112.9 ...

# multiple random data sets---------------------------------------------------
X <- mvnram(
  n = 100,
  A = A,
  S = S,
  F = F,
  I = I,
  mu = mu,
  R = 10
)
str(X, list.len = 6)

#> List of 10
#>  $ :'data.frame':	100 obs. of  3 variables:
#>   ..$ V1: num [1:100] 117.9 113.9 101.6 76.7 110 ...
#>   ..$ V2: num [1:100] 105.8 125.5 100.5 104.4 91.9 ...
#>   ..$ V3: num [1:100] 112.2 112.1 91.1 85.7 96.5 ...
#>  $ :'data.frame':	100 obs. of  3 variables:
#>   ..$ V1: num [1:100] 88.6 96.9 82.2 127.2 93.9 ...
#>   ..$ V2: num [1:100] 83.1 95.5 87.8 124 111.4 ...
#>   ..$ V3: num [1:100] 84.4 66.8 105.3 110.8 92.1 ...
#>  $ :'data.frame':	100 obs. of  3 variables:
#>   ..$ V1: num [1:100] 81.7 140.2 129.8 126.5 115.5 ...
#>   ..$ V2: num [1:100] 81.7 119.5 106.5 102.6 105.2 ...
#>   ..$ V3: num [1:100] 74.4 123.8 102 102.3 122.9 ...
#>  $ :'data.frame':	100 obs. of  3 variables:
#>   ..$ V1: num [1:100] 110.6 112 101.2 96.7 101.3 ...
#>   ..$ V2: num [1:100] 117.6 121.3 107.5 98.6 85.9 ...
#>   ..$ V3: num [1:100] 115.6 113.8 103.7 114.4 85.4 ...
#>  $ :'data.frame':	100 obs. of  3 variables:
#>   ..$ V1: num [1:100] 135.2 87 105.6 80.6 102.6 ...
#>   ..$ V2: num [1:100] 131.7 72.6 120.8 88.3 120.8 ...
#>   ..$ V3: num [1:100] 108 80.2 109.4 97.6 106.1 ...
#>  $ :'data.frame':	100 obs. of  3 variables:
#>   ..$ V1: num [1:100] 111.9 105.3 120.3 48 97.2 ...
#>   ..$ V2: num [1:100] 108 99.5 133.4 89.3 80.4 ...
#>   ..$ V3: num [1:100] 105.9 106.8 109.8 79.8 114.7 ...
#>   [list output truncated]