Generate Multivariate Normal Data \( \mathbf{X} \sim \mathcal{N}_{k} \left( \boldsymbol{\mu}, \boldsymbol{\Sigma} \right) %(\#eq:dist-X-mvn) \) from the Reticular Action Model (RAM) Matrices.

The \(\mathbf{S}\) matrix is derived from a vector of variances.

mvnramsigma2(
  n,
  A,
  sigma2,
  F,
  I,
  M = NULL,
  mu = NULL,
  tol = 1e-06,
  empirical = FALSE,
  R = NULL,
  par = FALSE,
  ncores = NULL,
  mc = TRUE,
  lb = FALSE,
  cl_eval = FALSE,
  cl_export = FALSE,
  cl_expr,
  cl_vars
)

Arguments

n
Integer. Sample size.

A
\(\mathbf{A}\) matrix. Asymmetric paths (single-headed arrows), such as regression coefficients and factor loadings.

sigma2
Numeric vector. Vector of variances \(\sigma^2\). The first element should be the variance of an exogenous variable

F
\(\mathbf{F}\) matrix. Filter matrix used to select the observed variables.

I
\(\mathbf{I}\) matrix. Identity matrix.

M
Numeric vector. Mean structure. Vector of means and intercepts.

mu
Numeric vector. Location parameter mean vector \(\boldsymbol{\mu}\) of length \(k\).

tol
Numeric. Tolerance (relative to largest variance) for numerical lack of positive-definiteness in Sigma.

empirical
Logical. If TRUE, mu and Sigma specify the empirical not population mean and covariance matrix.

R
Integer. Number of Monte Carlo replications. If R is not provided, the function produces a single random data set. If R is an integer greater than 1, (e.g., R = 10), the function produces multiple random data sets stored in each element of a list of length R. par and all succeeding arguments are only relevant when R > 1.

par
Logical. If TRUE, use multiple cores. If FALSE, use lapply().

ncores
Integer. Number of cores to use if par = TRUE. If unspecified, defaults to detectCores() - 1.

mc
Logical. If TRUE, use parallel::mclapply(). If FALSE, use parallel::parLapply() or parallel::parLapplyLB(). Ignored if par = FALSE.

lb
Logical. If TRUE use parallel::parLapplyLB(). If FALSE, use parallel::parLapply(). Ignored if par = FALSE and mc = TRUE.

cl_eval
Logical. Execute parallel::clusterEvalQ() using cl_expr. Ignored if mc = TRUE.

cl_export
Logical. Execute parallel::clusterExport() using cl_vars. Ignored if mc = TRUE.

cl_expr
Expression. Expression passed to parallel::clusterEvalQ() Ignored if mc = TRUE.

cl_vars
Character vector. Names of objects to pass to parallel::clusterExport() Ignored if mc = TRUE.

Value

If R = NULL or R = 1, returns an \(n \times k\) multivariate normal data matrix generated using the provided or derived mean vector mu \(\left( \boldsymbol{\mu} \right)\) and the derived variance-covariance matrix Sigma \(\left( \boldsymbol{\Sigma} \right)\). If R is an integer greater than 1, (e.g., R = 10) returns a list of length R of \(n \times k\) multivariate normal data matrices generated using the provided or derived mean vector mu and the derived variance-covariance matrix Sigma.

See also

Examples

# One-factor CFA model--------------------------------------------------------
A <- matrix(data = 0, ncol = 6, nrow = 6)
for (i in 2:6) {
  A[i, 1] <- 0.5
}
sigma2 <- c(1, 1, 1, 1, 1, 1)
F <- I <- diag(nrow(A))
F <- diag(nrow(A) - 1)
F <- cbind(0, F)
X <- mvnramsigma2(
  n = 100,
  A = A,
  sigma2 = sigma2,
  F = F,
  I = I,
  empirical = TRUE
)

#> mu = NULL and M = NULL. mu is set to a vector of zeroes of length 5.

str(X, list.len = 6)

#> 'data.frame':	100 obs. of  5 variables:
#>  $ V1: num  -0.1912 0.0694 1.5068 0.1594 0.887 ...
#>  $ V2: num  0.293 1.637 1.693 0.292 -0.323 ...
#>  $ V3: num  -0.046 -1.396 1.249 0.118 -0.747 ...
#>  $ V4: num  2.3601 -0.0078 -0.089 -0.2348 0.8273 ...
#>  $ V5: num  -0.414 0.953 1.095 0.269 0.316 ...

cov(X)

#>      V1   V2   V3   V4   V5
#> V1 1.00 0.25 0.25 0.25 0.25
#> V2 0.25 1.00 0.25 0.25 0.25
#> V3 0.25 0.25 1.00 0.25 0.25
#> V4 0.25 0.25 0.25 1.00 0.25
#> V5 0.25 0.25 0.25 0.25 1.00