Generate Univariate Data

Generates an \(n \times 1\) univariate data vector.

univ(n, rFUN = rnorm, R = NULL, par = FALSE, ncores = NULL, ...)

Arguments

n
Integer. Sample size.

rFUN
Function. Data generating function to generate univariate data.

R
Integer. Number of Monte Carlo replications. By default, R = NULL produces a single random data set. If R is defined (e.g., R = 10), produces a list of length R of multiple random data sets.

par
Logical. If TRUE, use multiple cores.

ncores
Integer. Number of cores to use if par = TRUE. If unspecified, defaults to detectCores() - 1.

...
Arguments to pass to rFUN.

Value

If R = NULL, returns an \(n \times 1\) univariate data vector generated using rFUN and parameters passed to .... If R is an integer, returns a list of length R of \(n \times 1\) univariate data vector generated using rFUN and parameters passed to ....

Details

The univariate distribution and parameters used in the data generating process can be specified using rFUN and ....

Examples

n <- 5
R <- 5
#################################
# normal distribution
#################################
mu <- 100
sigma2 <- 225
sigma <- sqrt(sigma2)
x <- univ(
  n = n,
  rFUN = rnorm,
  mean = mu,
  sd = sigma
)
str(x)

#>  num [1:5] 109.9 101.6 90 105.4 88.1

x <- univ(
  n = n,
  rFUN = rnorm,
  R = R,
  mean = mu,
  sd = sigma
)
str(x)

#> List of 5
#>  $ : num [1:5] 73.1 77.9 102.9 98.1 96
#>  $ : num [1:5] 96.6 129 93.8 88.4 124.6
#>  $ : num [1:5] 76.2 113.9 105.3 104.1 103.4
#>  $ : num [1:5] 84.3 111.6 100.4 117.3 103.8
#>  $ : num [1:5] 87.6 104.3 114.1 92 80.5

#################################
# binomial distribution
#################################
n_trials <- 1
p <- 0.50
x <- univ(
  n = n,
  rFUN = rbinom,
  size = n_trials,
  prob = p
)
str(x)

#>  int [1:5] 1 0 1 0 0

x <- univ(
  n = n,
  rFUN = rbinom,
  R = R,
  size = n_trials,
  prob = p
)
str(x)

#> List of 5
#>  $ : int [1:5] 1 0 0 0 0
#>  $ : int [1:5] 1 0 0 1 1
#>  $ : int [1:5] 1 0 0 0 1
#>  $ : int [1:5] 1 0 1 0 0
#>  $ : int [1:5] 1 0 0 1 1