Generate Multivariate Normal Data \(\mathbf{X} \sim \mathcal{N}_{k} \left( \boldsymbol{\mu}, \boldsymbol{\Sigma}\right)\)

Generates an \(n \times k\) multivariate normal data matrix \(\mathbf{X}\) from a \(k \times 1\) mean vector \(\boldsymbol{\mu}\) and a \(k \times k\) variance-covariance matrix \(\boldsymbol{\Sigma}\). This function is a wrapper around MASS::mvrnorm().

mvn(n, mu = NULL, Sigma, R = NULL, par = FALSE, ncores = NULL, ...)

Arguments

n
Integer. Sample size.

mu
Numeric vector. Location parameter mean vector of length \(k\) (\(\boldsymbol{\mu}\)).

Sigma
Numeric matrix. \(k \times k\) variance-covariance matrix (\(\boldsymbol{\Sigma}\)).

R
Integer. Number of Monte Carlo replications. By default, R = NULL produces a single random data set. If R is defined (e.g., R = 10), produces a list of length R of multiple random data sets.

par
Logical. If TRUE, use multiple cores.

ncores
Integer. Number of cores to use if par = TRUE. If unspecified, defaults to detectCores() - 1.

...
Arguments that can be passed to MASS::mvrnorm().

Value

If R = NULL, returns an \(n \times k\) multivariate normal data matrix generated using the mean vector mu \(\left( \boldsymbol{\mu} \right)\) and the variance-covariance matrix Sigma \(\left( \boldsymbol{\Sigma} \right)\) provided. If R is an integer, returns a list of length R of \(n \times k\) multivariate normal data matrices generated using the mean vector mu`` and the variance-covariance matrix Sigma` provided.

Details

The multivariate normal (or multivariate Gaussian, or joint normal) distribution is given by $$ f_{\mathbf{X}} \left( X_1, \dots, X_k \right) = \frac{ \exp \left[ - \frac{1}{2} \left( \mathbf{X} - \boldsymbol{\mu} \right)^{\prime} \boldsymbol{\Sigma}^{-1} \left( \mathbf{X} - \boldsymbol{\mu} \right) \right] }{ \sqrt{ \left( 2 \pi \right)^{k} | \boldsymbol{\Sigma} | } } $$ with \(k\)-dimensional random vector \(\mathbf{X} \in \boldsymbol{\mu} + \textrm{span}\left(\boldsymbol{\Sigma}\right) \subseteq \mathbf{R}^k\), \(\boldsymbol{\mu}\) is the location parameter mean (\(\boldsymbol{\mu} \in \mathbf{R}^k\)), and \(\boldsymbol{\Sigma}\) is the variance-covariance matrix (\(\boldsymbol{\Sigma} \in \mathbf{R}^{k \times k}\)).

References

Venables, W. N., Ripley, B. D., & Venables, W. N. (2002). Modern applied statistics with S. New York, N.Y: Springer.

Wikipedia: Multivariate normal distribution

See also

Examples

Sigma <- matrix(
  data = c(
    225, 112.50, 56.25,
    112.5, 225, 112.5,
    56.25, 112.50, 225
  ),
  ncol = 3
)
mu <- c(100, 100, 100)
# single random data set
X <- mvn(
  n = 100,
  mu = mu,
  Sigma = Sigma
)
str(X, list.len = 6)

#>  num [1:100, 1:3] 110 95.9 116.9 91.6 66.4 ...
#>  - attr(*, "dimnames")=List of 2
#>   ..$ : NULL
#>   ..$ : NULL

# multiple random data sets
X <- mvn(
  n = 100,
  mu = mu,
  Sigma = Sigma,
  R = 10
)
str(X, list.len = 6)

#> List of 10
#>  $ : num [1:100, 1:3] 116.8 106.9 80.5 97.2 78.8 ...
#>   ..- attr(*, "dimnames")=List of 2
#>   .. ..$ : NULL
#>   .. ..$ : NULL
#>  $ : num [1:100, 1:3] 112.3 87.8 98.7 81.8 102.3 ...
#>   ..- attr(*, "dimnames")=List of 2
#>   .. ..$ : NULL
#>   .. ..$ : NULL
#>  $ : num [1:100, 1:3] 74.3 75.8 92.1 88.4 89.3 ...
#>   ..- attr(*, "dimnames")=List of 2
#>   .. ..$ : NULL
#>   .. ..$ : NULL
#>  $ : num [1:100, 1:3] 94.5 62 112.1 108.2 94 ...
#>   ..- attr(*, "dimnames")=List of 2
#>   .. ..$ : NULL
#>   .. ..$ : NULL
#>  $ : num [1:100, 1:3] 102 123 129 127 118 ...
#>   ..- attr(*, "dimnames")=List of 2
#>   .. ..$ : NULL
#>   .. ..$ : NULL
#>  $ : num [1:100, 1:3] 81.7 82.2 79.7 109 87.6 ...
#>   ..- attr(*, "dimnames")=List of 2
#>   .. ..$ : NULL
#>   .. ..$ : NULL
#>   [list output truncated]