Generate Multivariate Normal Data from RAM Matrices

Generates multivariate normal data from the Reticular Action Model (RAM) matrices.

mvnram(
  n,
  A,
  S,
  F,
  I,
  M = NULL,
  mu = NULL,
  R = NULL,
  par = FALSE,
  ncores = NULL,
  ...
)

Arguments

n
Integer. Sample size.

A
Asymmetric paths, such as regression coefficients and factor loadings.

S
Symmetric matrix representing variances and covariances.

F
Filter matrix used to select the observed variables.

I
Identity matrix.

M
Vector of means and intercepts.

mu
Numeric vector. Location parameter mean vector of length \(k\) (\(\boldsymbol{\mu}\)).

R
Integer. Number of Monte Carlo replications. By default, R = NULL produces a single random data set. If R is defined (e.g., R = 10), produces a list of length R of multiple random data sets.

par
Logical. If TRUE, use multiple cores.

ncores
Integer. Number of cores to use if par = TRUE. If unspecified, defaults to detectCores() - 1.

...
Arguments that can be passed to MASS::mvrnorm().

Value

If R = NULL, returns an \(n \times k\) multivariate normal data matrix generated using the provided or derived mean vector mu \(\left( \boldsymbol{\mu} \right)\) and the derived variance-covariance matrix Sigma \(\left( \boldsymbol{\Sigma} \right)\). If R is an integer, returns a list of length R of \(n \times k\) multivariate normal data matrices generated using the provided or derived mean vector mu and the derived variance-covariance matrix Sigma.

Details

The multivariate normal distribution has two parameters, namely the \(k \times 1\) mean vector mu \(\left( \boldsymbol{\mu} \right)\) and the \(k \times k\) variance-covariance matrix Sigma \(\left( \boldsymbol{\Sigma} \right)\).

The mean vector mu \(\left( \boldsymbol{\mu} \right)\) can be supplied directly using the mu argument. It can also be derived using M . Note that the argument mu \(\left( \boldsymbol{\mu} \right)\) takes precedence over M If mu is not provided, it is computed using M with the jeksterslabRsem::ram_mutheta() function. If both mu and M are not provided, mu is set to a vector of zeroes with the appropriate length.

The variance-covariance matrix Sigma \(\left(\boldsymbol{\Sigma}\right)\) is derived from the RAM matrices A, S, F, and I.

mu \(\left( \boldsymbol{\mu} \right)\) and Sigma \(\left(\boldsymbol{\Sigma}\right)\) are then used by the mvn() to generate multivariate normal data.

References

McArdle, J. J. (2013). The development of the RAM rules for latent variable structural equation modeling. In A. Maydeu-Olivares & J. J. McArdle (Eds.), Contemporary Psychometrics: A festschrift for Roderick P. McDonald (pp. 225--273). Lawrence Erlbaum Associates.

McArdle, J. J., & McDonald, R. P. (1984). Some algebraic properties of the Reticular Action Model for moment structures. British Journal of Mathematical and Statistical Psychology, 37(2), 234--251.

See also

Examples

A <- matrix(
  data = c(
    0, 0.26^(1 / 2), 0,
    0, 0, 0.26^(1 / 2),
    0, 0, 0
  ),
  ncol = 3
)
S <- F <- I <- diag(3)
S[1, 1] <- 225
S[2, 2] <- 166.5
S[3, 3] <- 166.5
mu <- c(100, 100, 100)
# single random data set
X <- mvnram(
  n = 100,
  A = A,
  S = S,
  F = F,
  I = I,
  mu = mu
)
str(X, list.len = 6)

#>  num [1:100, 1:3] 94.8 105.7 99.3 92.8 93.5 ...
#>  - attr(*, "dimnames")=List of 2
#>   ..$ : NULL
#>   ..$ : NULL

# multiple random data sets
X <- mvnram(
  n = 100,
  A = A,
  S = S,
  F = F,
  I = I,
  mu = mu,
  R = 10
)
str(X, list.len = 6)

#> List of 10
#>  $ : num [1:100, 1:3] 107.9 107.4 109.5 108.3 94.4 ...
#>   ..- attr(*, "dimnames")=List of 2
#>   .. ..$ : NULL
#>   .. ..$ : NULL
#>  $ : num [1:100, 1:3] 108.9 115.5 88.8 118.6 114.2 ...
#>   ..- attr(*, "dimnames")=List of 2
#>   .. ..$ : NULL
#>   .. ..$ : NULL
#>  $ : num [1:100, 1:3] 132.4 112.9 76.7 115.7 65.9 ...
#>   ..- attr(*, "dimnames")=List of 2
#>   .. ..$ : NULL
#>   .. ..$ : NULL
#>  $ : num [1:100, 1:3] 84.2 117 103.2 101.8 112.2 ...
#>   ..- attr(*, "dimnames")=List of 2
#>   .. ..$ : NULL
#>   .. ..$ : NULL
#>  $ : num [1:100, 1:3] 139.2 103.6 104 95.9 103.8 ...
#>   ..- attr(*, "dimnames")=List of 2
#>   .. ..$ : NULL
#>   .. ..$ : NULL
#>  $ : num [1:100, 1:3] 108.3 99.9 107.8 110.8 129.2 ...
#>   ..- attr(*, "dimnames")=List of 2
#>   .. ..$ : NULL
#>   .. ..$ : NULL
#>   [list output truncated]