Generate Univariate Data

Generates an \(n \times 1\) univariate data vector or a list of \(n \times 1\) univariate data vectors of length R. The default data generating function is the normal distribution $$ X \sim \mathcal{N} \left( \mu, \sigma^2 \right) . %(\#eq:dist-X-norm) $$

univ(
  n,
  rFUN = rnorm,
  ...,
  R = NULL,
  par = FALSE,
  ncores = NULL,
  mc = TRUE,
  lb = FALSE,
  cl_eval = FALSE,
  cl_export = FALSE,
  cl_expr,
  cl_vars,
  rbind = NULL
)

Arguments

n
Integer. Sample size.

rFUN
Function. Data generating function to generate univariate data.

...
Arguments to pass to rFUN.

R
Integer. Number of Monte Carlo replications. If R is not provided, the function produces a single random data set. If R is an integer greater than 1, (e.g., R = 10), the function produces multiple random data sets stored in each element of a list of length R. par and all succeeding arguments are only relevant when R > 1.

par
Logical. If TRUE, use multiple cores. If FALSE, use lapply().

ncores
Integer. Number of cores to use if par = TRUE. If unspecified, defaults to detectCores() - 1.

mc
Logical. If TRUE, use parallel::mclapply(). If FALSE, use parallel::parLapply() or parallel::parLapplyLB(). Ignored if par = FALSE.

lb
Logical. If TRUE use parallel::parLapplyLB(). If FALSE, use parallel::parLapply(). Ignored if par = FALSE and mc = TRUE.

cl_eval
Logical. Execute parallel::clusterEvalQ() using cl_expr. Ignored if mc = TRUE.

cl_export
Logical. Execute parallel::clusterExport() using cl_vars. Ignored if mc = TRUE.

cl_expr
Expression. Expression passed to parallel::clusterEvalQ() Ignored if mc = TRUE.

cl_vars
Character vector. Names of objects to pass to parallel::clusterExport() Ignored if mc = TRUE.

rbind
NULL or logical. If rbind = NULL, returns the list produced. If TRUE, uses rbind() to bind the rows of the list produced. If FALSE, uses cbind() to bind the columns of the list produced. Test that each element of the output list has the appropriate dimensions for binding before using this option.

Value

If R = NULL or R = 1, returns an \( n \times 1 \) univariate data vector generated using rFUN and parameters passed to .... If R is an integer greater than 1, (e.g., R = 10) returns a list of length R of \( n \times 1 \) univariate data vector generated using rFUN and parameters passed to ....

Details

The univariate distribution and parameters used in the data generating process can be specified using rFUN and ....

Options for explicit parallelism are provided when R > 1 especially when R is large. See par and suceeding arguments.

Examples

x <- univ(n = 100, rFUN = rnorm, mean = 100, sd = sqrt(225))
str(x)

#>  num [1:100] 98.6 87.6 99.7 104.8 117.6 ...

xstar <- univ(n = 100, rFUN = rnorm, mean = 100, sd = sqrt(225), R = 100)
str(xstar, list.len = 6)

#> List of 100
#>  $ : num [1:100] 107.1 85.4 105.4 98.8 86.4 ...
#>  $ : num [1:100] 96.2 78 83.4 109.2 137.5 ...
#>  $ : num [1:100] 113.7 90.9 90 112 82.7 ...
#>  $ : num [1:100] 71.6 86.6 107.6 97 130.7 ...
#>  $ : num [1:100] 96 110.4 105.6 116.3 99.3 ...
#>  $ : num [1:100] 102.7 82.6 72.2 90.6 90.2 ...
#>   [list output truncated]